+++ apps/io-wallet-user-func/src/infra/attestation-service/android/assertion.ts
@@ -52,6 +52,8 @@ export const verifyAssertion = async (params: VerifyAssertionParams) => {
     integrityAssertion,
   } = params;
 
+  const errors: string[] = [];
+
   // First check whether the clientData has been signed correctly with the hardware key
   const signatureValidated = validateAssertionSignature(
     hardwareKey,
@@ -100,23 +102,26 @@ export const verifyAssertion = async (params: VerifyAssertionParams) => {
           androidPlayStoreCertificateHash,
         );
         break;
-      } catch (e) {
+      } catch (error) {
         /* If it fails I continue the for loop anyway to try other bundleIdentifiers.
          * The check is still done at the end on the value of responseValidated
          */
+        errors.push(`${error}`);
       }
+    } else {
+      errors.push(`${result}`);
     }
   }
 
   if (!tokenPayloadExternal || !bundleIdentifier) {
     throw new Error(
-      "[Android Assertion] Invalid token payload from Play Integrity API response",
+      `[Android Assertion] Invalid token payload from Play Integrity API response: ${errors.join(",")}`,
     );
   }
 
   if (!responseValidated) {
     throw new Error(
-      "[Android Assertion] Integrity Response did not pass validation",
+      `[Android Assertion] Integrity Response did not pass validation: ${errors.join(",")}`,
     );
   }
 };