+++ apps/io-wallet-user-func/src/infra/attestation-service/android/attestation.ts
@@ -132,11 +132,25 @@ export const validateRevocation = async (
     });
   }
   const crl = (await res.json()) as CRL; // Add type assertion for crl
-  const isRevoked = x509Chain.some((cert) => cert.serialNumber in crl.entries);
+
+  const revokedSerials = Object.keys(crl.entries).map((key) =>
+    key.toLowerCase(),
+  );
+
+  const isRevoked = x509Chain.some((cert) => {
+    const currentSn = cert.serialNumber.toLowerCase();
+    return revokedSerials.some((revokedSerial) =>
+      currentSn.includes(revokedSerial),
+    );
+  });
+
   if (isRevoked) {
-    throw new AndroidAttestationError("Certificate is revoked", {
-      x509Chain,
-    });
+    throw new AndroidAttestationError(
+      "A certificate within the chain has been revoked by Google",
+      {
+        x509Chain,
+      },
+    );
   }
   return x509Chain;
 };