+++ apps/io-wallet-user-func/src/infra/pid-issuer/client.ts
@@ -59,7 +59,6 @@ export class PidIssuerClient
             // the endpoint returns 404 if the credentials have already been revoked or do not exist
             // the credentials already revoked => the error is suppressed because our endpoint is idempotent
             // the credentials do not exist => the error is suppressed because our endpoint returns 204 when attempting to revoke non-existent credentials
-            // TODO: SIW-1402 add log
             return undefined;
           }
           if (!result.ok) {