+++ apps/io-wallet-user-func/CHANGELOG.md
@@ -1,5 +1,31 @@
 # io-wallet-user-func
 
+## 1.3.6
+
+### Patch Changes
+
+- 35fdcb0: Moved IO Web token validation to APIM
+- Updated dependencies [35fdcb0]
+  - io-wallet-common@1.1.1
+
+## 1.3.5
+
+### Patch Changes
+
+- 47bcac6: Fix Android Attestation revocation validation
+
+## 1.3.4
+
+### Patch Changes
+
+- 9486e3c: Added `requests` telemetry in Application Insights
+
+## 1.3.3
+
+### Patch Changes
+
+- f9dbfdd: Add integrity error logs
+
 ## 1.3.2
 
 ### Patch Changes